Add markdown as rendering type
