+++ src/components/interactive-builder/modals/question/question-form/rendering-types/inputs/index.tsx
@@ -4,3 +4,4 @@ export { default as Text } from './text/text.component';
 export { default as TextArea } from './text-area/textarea.component';
 export { default as Toggle } from './toggle/toggle.component';
 export { default as UiSelectExtended } from './ui-select-extended/ui-select-extended.component';
+export { default as Markdown } from './markdown/markdown.component';
