+++ src/components/interactive-builder/modals/question/question-form/rendering-types/rendering-type.component.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Date, Number, Text, TextArea, Toggle, UiSelectExtended } from './inputs';
+import { Date, Markdown, Number, Text, TextArea, Toggle, UiSelectExtended } from './inputs';
 import type { RenderType } from '@openmrs/esm-form-engine-lib';
 import type { ComponentProps } from '@types';
 import { renderTypeOptions, renderingTypes } from '@constants';
@@ -12,6 +12,7 @@ const componentMap: Partial<Record<RenderType, React.FC<ComponentProps>>> = {
   'ui-select-extended': UiSelectExtended,
   date: Date,
   datetime: Date,
+  markdown: Markdown,
 };
 
 const RenderTypeComponent: React.FC<ComponentProps> = ({ formField, setFormField }) => {
